refactor(mediator): rename Mediator to DynamicDispatchMediator
